// lib/interfaces/i-header-config.ts
/**
 * @author Oguntuberu Nathan O. <nateoguns.work@gmail.com>
**/

export default interface HeaderConfiguration {
    oauth_callback?: string;
    oauth_consumer_key?: string;
    oauth_consumer_secret?: string;
    oauth_nonce?: string;
    oauth_signature_method?: string;
    oauth_timestamp?: string;
    oauth_token?: string;
    oauth_token_secret?: string;
    oauth_version?: string;
}